removal buffer added to delay re-adding symbols
When a symbol is removed, its Ticker data may still be
streamed from WSS for a short amount of time. To prevent
these Ticker objects being rendered, the removal buffer
caches the removed symbol and the time it was removed.
If a Ticker arrives from WSS whose symbol is in the
remove buffer, it won't be rendered. The symbol is deleted
from the remove buffer after 10 seconds. When it's
deleted, the associated symbol can be re-added again.